refactor: contextprovider 제거
